Changes from Western New England Comp

+++ src/main/java/frc/robot/commands/intakerelease/ReleaseCMD.java
@@ -29,7 +29,9 @@ public class ReleaseCMD extends CommandBase {
   public void execute() {
     if ((!arm.hasDropped())
         && intakerelease.haveCone()
+        && intakerelease.wantsCone()
         && (arm.getExtention() >= 0.2 && arm.getRotation() >= 0.15)
+        && !timer.hasElapsed(1.0)
         && needsDrop) {
       arm.drop();
       return;
@@ -43,6 +45,6 @@ public class ReleaseCMD extends CommandBase {
 
   @Override
   public boolean isFinished() {
-    return !intakerelease.haveCone() && !intakerelease.haveCube() && timer.hasElapsed(1.0);
+    return timer.hasElapsed(1.5);
   }
 }
